+++ list_picker/ui/keys.py
@@ -1,3 +1,4 @@
+#!/bin/python
 import curses
 """
 Define key dictionary for controlling list_picker.
@@ -15,8 +16,8 @@ list_picker_keys = {
     "cursor_up":                        [ord('k'), curses.KEY_UP],
     "half_page_up":                     [ord('u')],
     "half_page_down":                   [ord('d')],
-    "page_up":                          [ord('p'), curses.KEY_PPAGE, 2],
-    "page_down":                        [ord('n'), curses.KEY_NPAGE, 6],
+    "page_up":                          [curses.KEY_PPAGE, 2], # Ctrl+b
+    "page_down":                        [curses.KEY_NPAGE, 6], # Ctrl+f
     "cursor_bottom":                    [ord('G'), curses.KEY_END],
     "cursor_top":                       [ord('g'), curses.KEY_HOME],
     "five_up":                          [ord('K')],
@@ -40,8 +41,8 @@ list_picker_keys = {
     "search_input":                     [ord('/')],
     "settings_input":                   [ord('`')],
     "settings_options":                 [ord('~')],
-    "continue_search_forward":          [ord('i')],
-    "continue_search_backward":         [ord('I')],
+    "continue_search_forward":          [ord('n'), ord('i')],
+    "continue_search_backward":         [ord('N'), ord('i')],
     "cancel":                           [27], # Escape key
     "opts_input":                       [ord(':')],
     "opts_select":                      [ord('o')],
@@ -72,8 +73,8 @@ help_keys = {
     "cursor_up":                        [ord('k'), curses.KEY_UP],
     "half_page_up":                     [ord('u')],
     "half_page_down":                   [ord('d')],
-    "page_up":                          [ord('p'), curses.KEY_PPAGE, 2],
-    "page_down":                        [ord('n'), curses.KEY_NPAGE, 6],
+    "page_up":                          [curses.KEY_PPAGE, 2],
+    "page_down":                        [curses.KEY_NPAGE, 6],
     "cursor_bottom":                    [ord('G'), curses.KEY_END],
     "cursor_top":                       [ord('g'), curses.KEY_HOME],
     "five_up":                          [ord('K')],
@@ -88,8 +89,8 @@ help_keys = {
     "search_input":                     [ord('/')],
     "settings_input":                   [ord('`')],
     "settings_options":                 [ord('~')],
-    "continue_search_forward":          [ord('i')],
-    "continue_search_backward":         [ord('I')],
+    "continue_search_forward":          [ord('n'), ord('i')],
+    "continue_search_backward":         [ord('N'), ord('i')],
     "cancel":                           [27], # Escape key
     "col_select":                       [ord('0'), ord('1'), ord('2'), ord('3'), ord('4'), ord('5'), ord('6'), ord('7'), ord('8'), ord('9')],
     "col_select_next":                             [ord('>')], 
@@ -107,8 +108,8 @@ notification_keys = {
     "cursor_up":                        [ord('k'), curses.KEY_UP],
     "half_page_up":                     [ord('u')],
     "half_page_down":                   [ord('d')],
-    "page_up":                          [ord('p'), curses.KEY_PPAGE],
-    "page_down":                        [ord('n'), curses.KEY_NPAGE],
+    "page_up":                          [curses.KEY_PPAGE, 2],
+    "page_down":                        [curses.KEY_NPAGE, 6],
     "cursor_bottom":                    [ord('G'), curses.KEY_END],
     "cursor_top":                       [ord('g'), curses.KEY_HOME],
     "five_up":                          [ord('K')],
@@ -125,8 +126,8 @@ menu_keys = {
     "cursor_up":                        [ord('k'), curses.KEY_UP],
     "half_page_up":                     [ord('u')],
     "half_page_down":                   [ord('d')],
-    "page_up":                          [ord('p'), curses.KEY_PPAGE, 2],
-    "page_down":                        [ord('n'), curses.KEY_NPAGE, 6],
+    "page_up":                          [curses.KEY_PPAGE, 2],
+    "page_down":                        [curses.KEY_NPAGE, 6],
     "cursor_bottom":                    [ord('G'), curses.KEY_END],
     "cursor_top":                       [ord('g'), curses.KEY_HOME],
     "five_up":                          [ord('K')],
@@ -135,9 +136,9 @@ menu_keys = {
     "redraw_screen":                    [12], # Ctrl-l
     "filter_input":                     [ord('f')],
     "search_input":                     [ord('/')],
-    "continue_search_forward":          [ord('i')],
+    "continue_search_forward":          [ord('n'), ord('i')],
+    "continue_search_backward":         [ord('N'), ord('i')],
     "cancel":                           [27], # Escape key
-    "continue_search_backward":         [ord('I')],
     "opts_input":                       [ord(':')],
     "mode_next":                        [9], # Tab key
     "mode_prev":                        [353], # Shift+Tab key
@@ -151,8 +152,8 @@ options_keys = {
     "cursor_up":                        [ord('k'), curses.KEY_UP],
     "half_page_up":                     [ord('u')],
     "half_page_down":                   [ord('d')],
-    "page_up":                          [ord('p'), curses.KEY_PPAGE, 2],
-    "page_down":                        [ord('n'), curses.KEY_NPAGE, 6],
+    "page_up":                          [curses.KEY_PPAGE, 2],
+    "page_down":                        [curses.KEY_NPAGE, 6],
     "cursor_bottom":                    [ord('G'), curses.KEY_END],
     "cursor_top":                       [ord('g'), curses.KEY_HOME],
     "five_up":                          [ord('K')],
@@ -170,8 +171,8 @@ options_keys = {
     "filter_input":                     [ord('f')],
     "search_input":                     [ord('/')],
     "settings_input":                   [ord('`')],
-    "continue_search_forward":          [ord('i')],
-    "continue_search_backward":         [ord('I')],
+    "continue_search_forward":          [ord('n'), ord('i')],
+    "continue_search_backward":         [ord('N'), ord('i')],
     "cancel":                           [27], # Escape key
     "col_select":                       [ord('0'), ord('1'), ord('2'), ord('3'), ord('4'), ord('5'), ord('6'), ord('7'), ord('8'), ord('9')],
     "refresh":                          [curses.KEY_F5, curses.KEY_RESIZE],