+++ list_picker/utils/clipboard_operations.py
@@ -1,3 +1,4 @@
+#!/bin/python
 import pyperclip
 from list_picker.utils.utils import get_selected_indices, format_row, format_row_full
 from typing import Tuple