+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG.md
 
+## [0.1.2] 2025-06-19
+ - Fixed bug where list_picker crashed when rapidly resizing terminal or rapidly changing font-size.
+ - Fixed bug with filtering/searching where multiple tokens could not be specified for the same column.
+ - Visual improvements:
+   - Changed the footer colour to match the title bar in the main theme.
+   - Right aligned the elements in the footer
+   - Improved the appearance of the refresh indicator.
+ - Pickle files can now be loaded from the command line in addition to being able to be loaded wile running the Picker.
+
 ## [0.1.2] 2025-06-18
  - Added the ability to edit current instance of Picker in ipython when Ctrl+e is pressed.
  - Quick-toggle footer with '_'.